refactor(cli): split cli.py into focused modules
Addresses #81. Extracts 1900-line cli.py into focused modules:

- `analyzers/discovery.py` — `discover_analyzers()`
- `utils/process.py` — `run_cmd`, `require_tool`
- `utils/report.py` — report helpers (`format_output_path`, `write_report`,
  `run_playbooks`, `validate_report`, `render_and_save_reports`, etc.)
- `commands/analyze.py` — `analyze`, `evaluate`, `list` commands
- `commands/bootstrap.py` — `bootstrap playbook/archive` commands
- `commands/archive.py` — `archive add` command
- `commands/rules.py` — `rules list/show/evaluate` commands
- `commands/check.py` — `check`, `version` commands

`cli.py` is now a ~65-line thin entry point that wires commands to `main`.
Backward-compat aliases (`_discover_analyzers`, `_run_cmd`, etc.) are
re-exported from `cli.py` to avoid breaking external references.

Updates `rules/evaluator.py` and all test mock patch targets to point to the
new module locations.
